actualizacion de jerarquías de proyecto
